email-worker/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
aiosmtplib==3.0.1
email-validator==2.0.0